+++ src/main/java/com/ua/project/Autobase/model/Application.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class Application {
     private Long id;
     private double weight;
-    private int cargoTypeId;
+    private long cargoTypeId;
 }
